requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.4
cmake==3.26.4
ffmpeg-python==0.2.0
filelock==3.12.2
frozenlist==1.3.3
future==0.18.3
h11==0.14.0
idna==3.4
Jinja2==3.1.2
lit==16.0.6
llvmlite==0.40.1
MarkupSafe==2.1.3
more-itertools==9.1.0
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numba==0.57.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openai-whisper==20230314
python-dotenv==1.0.0
regex==2023.6.3
requests==2.31.0
sniffio==1.3.0
starlette==0.28.0
sympy==1.12
tiktoken==0.3.1
torch==2.0.1
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.7.1
urllib3==2.0.3
uvicorn==0.22.0
yarl==1.9.2
